+++ R/oa2df.R
@@ -1,12 +1,12 @@
 #' Convert OpenAlex collection from list to data frame
 #'
 #' It converts bibliographic collections gathered from OpenAlex database \href{https://openalex.org/}{https://openalex.org/} into data frame.
-#' The function converts a collection of records about works, authors, institutions, venues or concepts obtained using \code{oa_request} into a data frame/tibble.
+#' The function converts a collection of records about works, authors, institutions, venues or keywords obtained using \code{oa_request} into a data frame/tibble.
 #'
 #' @param data List. Output of \code{oa_request}.
 #' @param entity Character. Scholarly entity of the search.
 #' The argument can be one of
-#' c("works", "authors", "institutions", "concepts", "funders", "sources", "publishers", "topics").
+#' c("works", "authors", "institutions", "keywords", "funders", "sources", "publishers", "topics").
 #' @param abstract Logical. If TRUE, the function returns also the abstract of each item.
 #' Ignored if entity is different from "works". Defaults to TRUE.
 #' @param verbose Logical.
@@ -48,6 +48,17 @@
 #'
 #' @export
 oa2df <- function(data, entity, options = NULL, count_only = FALSE, group_by = NULL, abstract = TRUE, verbose = TRUE) {
+
+  if (rlang::is_interactive()) {
+    rlang::warn(
+      "Note: `oa_fetch` and `oa2df` now return new names for some columns in openalexR v2.0.0.
+      See NEWS.md for the list of changes.
+      Call `get_coverage()` to view the all updated columns and their original names in OpenAlex.",
+      .frequency = "regularly",
+      .frequency_id = "oa2df_column_change"
+    )
+  }
+
   if (length(data) == 0) {
     return(NULL)
   }
@@ -72,6 +83,7 @@ oa2df <- function(data, entity, options = NULL, count_only = FALSE, group_by = N
     authors = authors2df(data, verbose),
     institutions = institutions2df(data, verbose),
     concepts = concepts2df(data, verbose),
+    keywords = keywords2df(data, verbose),
     funders = funders2df(data, verbose),
     sources = sources2df(data, verbose),
     publishers = publishers2df(data, verbose),
@@ -137,14 +149,16 @@ oa2df <- function(data, entity, options = NULL, count_only = FALSE, group_by = N
 works2df <- function(data, abstract = TRUE, verbose = TRUE,
                      pb = if (verbose) oa_progress(length(data)) else NULL) {
   col_order <- c(
-    "id", "title", "display_name", "author", "ab", "publication_date", "relevance_score",
-    "so", "so_id", "host_organization", "issn_l", "url", "pdf_url",
-    "license", "version", "first_page", "last_page", "volume", "issue", "is_oa",
-    "is_oa_anywhere", "oa_status", "oa_url", "any_repository_has_fulltext",
-    "language", "grants", "cited_by_count", "counts_by_year",
-    "publication_year", "cited_by_api_url", "ids", "doi", "type",
-    "referenced_works", "related_works", "is_paratext", "is_retracted",
-    "concepts", "topics"
+    "id", "title", "display_name", "authorships", "abstract", "doi",
+    "publication_date", "publication_year", "relevance_score", "fwci",
+    "cited_by_count", "counts_by_year", "cited_by_api_url", "ids", "type",
+    "is_oa", "is_oa_anywhere", "oa_status", "oa_url",
+    "any_repository_has_fulltext", "source_display_name", "source_id", "issn_l",
+    "host_organization", "host_organization_name",
+    "landing_page_url", "pdf_url", "license", "version", "referenced_works",
+    "referenced_works_count", "related_works", "concepts", "topics", "keywords",
+    "is_paratext", "is_retracted", "language", "grants", "apc",
+    "first_page", "last_page", "volume", "issue"
   )
   works_process <- tibble::tribble(
     ~type, ~field,
@@ -161,29 +175,26 @@ works2df <- function(data, abstract = TRUE, verbose = TRUE,
     "identical", "is_retracted",
     "identical", "relevance_score",
     "identical", "language",
+    "identical", "fwci",
+    "identical", "referenced_works_count",
     "flat", "grants",
     "flat", "referenced_works",
     "flat", "related_works",
+    "rbind_df", "keywords",
     "rbind_df", "counts_by_year",
     "rbind_df", "concepts",
+    "flat", "apc_list",
+    "flat", "apc_paid",
     "flat", "ids"
   )
 
-  venue_cols <- c(
-    url = "landing_page_url",
-    pdf_url = "pdf_url",
-    is_oa = "is_oa",
-    license = "license",
-    version = "version"
-  )
   so_cols <- c(
-    so_id = "id",
-    so = "display_name",
+    source_id = "id",
+    source_display_name = "display_name",
     issn_l = "issn_l",
-    host_organization = "host_organization_name"
+    host_organization = "host_organization",
+    host_organization_name = "host_organization_name"
   )
-  inst_cols <- c("id", "display_name", "ror", "country_code", "type", "lineage")
-  empty_inst <- empty_list(inst_cols)
 
   n <- length(data)
   list_df <- vector(mode = "list", length = n)
@@ -200,84 +211,47 @@ works2df <- function(data, abstract = TRUE, verbose = TRUE,
       fields$type,
       SIMPLIFY = FALSE
     )
-
-    author <- venue <- ab <- NULL
-
-    if (!is.null(paper$primary_location)) {
-      so_info <- paper$primary_location["source"]
-      so_info <- if (length(so_info[[1]]) == 0) NA else so_info[[1]]
-      venue_info <- replace_w_na(paper$primary_location[venue_cols])
-      venue <- setNames(
-        c(venue_info, so_info[so_cols]),
-        c(names(venue_cols), names(so_cols))
-      )
+    if (!is.null(sim_fields$publication_date)) {
+      sim_fields$publication_date <- as.Date(sim_fields$publication_date)
     }
-
-    # authorships and affilitation
-    if (!is.null(paper$authorships)) {
-      author <- subs_na(
-        lapply(paper$authorships, function(l) {
-          l_inst <- l$institutions
-          inst_idx <- lengths(l_inst) > 0
-          if (length(inst_idx) > 0 && any(inst_idx)) {
-            first_inst <- l_inst[inst_idx][[1]]
-            first_inst$lineage <- paste(first_inst$lineage, collapse = ", ")
-          } else {
-            first_inst <- empty_inst
-          }
-          first_inst <- prepend(first_inst, "institution")
-          aff_raw <- list(
-            au_affiliation_raw =
-              if (length(l$raw_affiliation_strings)) {
-                l$raw_affiliation_strings[[1]]
-              } else {
-                NA_character_
-              }
-          )
-          l_author <- if (length(l$author) > 0) {
-            prepend(replace_w_na(l$author), "au")
-          } else {
-            empty_list(c("au_id", "au_display_name", "au_orcid"))
-          }
-          c(l_author, l[c("author_position", "is_corresponding")], aff_raw, first_inst)
-        }), "rbind_df"
-      )
-    }
-
-    # Abstract
-    if (!is.null(paper$abstract_inverted_index) && abstract) {
-      ab <- abstract_build(paper$abstract_inverted_index)
-    }
+    authorships <- process_paper_authors(paper$authorships)
+    ab <- abstract_build(paper$abstract_inverted_index, abstract)
     paper_biblio <- replace_w_na(paper$biblio)
     open_access <- replace_w_na(paper$open_access)
     if (length(open_access) > 0) {
       names(open_access)[[1]] <- "is_oa_anywhere"
     }
 
+    so_info <- paper$primary_location
+    venue <- so_info[names(so_info) != "source"]
+    source <- so_info$source
+    if (!is.null(source)){
+      source <- setNames(source[so_cols], names(so_cols))
+    }
+
+    # Process APC
+    apc <- NULL
+    if (any(lengths(paper[c("apc_list", "apc_paid")]) > 0)) {
+      apc_fields <- list(
+        value = NA_real_, currency = NA_character_,
+        value_usd = NA_real_, provenance = NA_character_
+      )
+      apc_list <- paper$apc_list[lengths(paper$apc_list) != 0]
+      apc_paid <- paper$apc_paid[lengths(paper$apc_paid) != 0]
+      apc <- list(rbind.data.frame(
+        c(type = "list", utils::modifyList(apc_fields, as.list(apc_list))),
+        c(type = "paid", utils::modifyList(apc_fields, as.list(apc_paid)))
+      ))
+    }
     topics <- process_topics(paper, "score")
-    out_ls <- c(
-      sim_fields, venue, open_access, paper_biblio,
-      list(author = author, ab = ab), topics
-    )
+    out_ls <- c(sim_fields, venue, source, open_access, paper_biblio,
+                list(authorships = authorships, abstract = ab, apc = apc), topics)
     out_ls[sapply(out_ls, is.null)] <- NULL
     list_df[[i]] <- out_ls
   }
 
   out_df <- rbind_oa_ls(list_df)
   out_df[, intersect(col_order, names(out_df))]
-}
-
-abstract_build <- function(ab) {
-  if (is.null(ab)) {
-    return(NA)
-  }
-  w <- rep(names(ab), lengths(ab))
-  ind <- unlist(ab)
-  if (is.null(ind)) {
-    return("")
-  }
-
-  paste(w[order(ind)], collapse = " ", sep = "")
 }
 
 
@@ -302,17 +276,11 @@ abstract_build <- function(ab) {
 #' # University of Naples Federico II is associated to the OpenAlex id I71267560.
 #'
 #'
-#' query_author <- oa_query(
-#'   identifier = NULL,
+#' res <- oa_fetch(
 #'   entity = "authors",
 #'   last_known_institutions.id = "I71267560",
-#'   works_count = ">500"
-#' )
-#'
-#' res <- oa_request(
-#'   query_url = query_author,
-#'   count_only = FALSE,
-#'   verbose = FALSE
+#'   works_count = ">700",
+#'   output = "list"
 #' )
 #'
 #' df <- oa2df(res, entity = "authors")
@@ -325,10 +293,6 @@ authors2df <- function(data, verbose = TRUE,
                        pb = if (verbose) oa_progress(length(data)) else NULL) {
   n <- length(data)
   list_df <- vector(mode = "list", length = n)
-
-  inst_cols <- c("id", "display_name", "ror", "country_code", "type", "lineage")
-  empty_inst <- empty_list(inst_cols)
-  empty_inst$affiliations_other <- list(NULL)
 
   author_process <- tibble::tribble(
     ~type, ~field,
@@ -356,35 +320,24 @@ authors2df <- function(data, verbose = TRUE,
       fields$type,
       SIMPLIFY = FALSE
     )
-    sub_affiliation <- item$last_known_institutions
-    if (!is.null(sub_affiliation) && length(sub_affiliation)) {
-      sub_affiliation <- sub_affiliation[[1]]
-      if (is.na(sub_affiliation[[1]])) {
-        sub_affiliation <- empty_inst
-      }
-      sub_affiliation$lineage <- paste(sub_affiliation$lineage, collapse = ", ")
-      sub_affiliation <- prepend(sub_affiliation, "affiliation")
-    }
-    sub_affiliation <- replace_w_na(sub_affiliation)
 
-    if (!is.null(item$affiliations)) {
-      affiliations_other <- sapply(item$affiliations, function(x) x$institution$id)
-      if (!is.null(sub_affiliation$affiliation_id)) {
-        affiliations_other <- affiliations_other[affiliations_other != sub_affiliation$affiliation_id]
-      }
-      sub_affiliation$affiliations_other <- list(affiliations_other)
+    # current affiliations
+    if (!is.null(item$last_known_institutions)) {
+      l_inst <- item$last_known_institutions
+      affs <- list(last_known_institutions = process_affil(l_inst))
+    } else {
+      affs <- NULL
     }
+
     topics <- process_topics(item, "count")
-    list_df[[i]] <- c(sim_fields, sub_affiliation, topics)
+    list_df[[i]] <- c(sim_fields, affs, item$summary_stats, topics)
   }
 
   col_order <- c(
     "id", "display_name", "display_name_alternatives", "relevance_score",
     "ids", "orcid", "works_count", "cited_by_count", "counts_by_year",
-    "affiliation_display_name", "affiliation_id", "affiliation_ror",
-    "affiliation_country_code", "affiliation_type", "affiliation_lineage",
-    "affiliations_other",
-    "topics", "works_api_url"
+    "2yr_mean_citedness", "h_index", "i10_index",
+    "last_known_institutions", "topics", "works_api_url"
   )
 
   out_df <- rbind_oa_ls(list_df)
@@ -409,16 +362,11 @@ authors2df <- function(data, verbose = TRUE,
 #'
 #' # Query to search information about all Italian educational institutions
 #'
-#' query_inst <- oa_query(
+#' res <- oa_fetch(
 #'   entity = "institutions",
 #'   country_code = "it",
-#'   type = "education"
-#' )
-#'
-#' res <- oa_request(
-#'   query_url = query_inst,
-#'   count_only = FALSE,
-#'   verbose = FALSE
+#'   type = "education",
+#'   output = "list"
 #' )
 #'
 #' oa2df(res, entity = "institutions")
@@ -446,6 +394,7 @@ institutions2df <- function(data, verbose = TRUE,
     "identical", "updated_date",
     "identical", "created_date",
     "identical", "relevance_score",
+    "flat", "summary_stats",
     "flat", "display_name_alternatives",
     "flat", "display_name_acronyms",
     "row_df", "geo",
@@ -468,7 +417,7 @@ institutions2df <- function(data, verbose = TRUE,
     interna <- NULL
     if (!is.null(item$international)) {
       interna <- list(
-        display_name_international = subs_na(
+        international_display_name = subs_na(
           item$international$display_name,
           type = "flat"
         )
@@ -481,10 +430,10 @@ institutions2df <- function(data, verbose = TRUE,
 
   col_order <- c(
     "id", "display_name", "display_name_alternatives", "display_name_acronyms",
-    "display_name_international", "ror", "ids", "country_code", "geo", "type",
+    "international_display_name", "ror", "ids", "country_code", "geo", "type",
     "homepage_url", "image_url", "image_thumbnail_url",
     "associated_institutions", "relevance_score", "works_count",
-    "cited_by_count", "counts_by_year",
+    "cited_by_count", "counts_by_year", "summary_stats",
     "works_api_url", "topics", "updated_date", "created_date"
   )
 
@@ -511,15 +460,10 @@ institutions2df <- function(data, verbose = TRUE,
 #' # Query to search information about all Italian educational institutions
 #'
 #'
-#' query_inst <- oa_query(
+#' res <- oa_query(
 #'   entity = "concepts",
-#'   display_name.search = "electrodynamics"
-#' )
-#'
-#' res <- oa_request(
-#'   query_url = query_inst,
-#'   count_only = FALSE,
-#'   verbose = FALSE
+#'   display_name.search = "electrodynamics",
+#'   output = "list"
 #' )
 #'
 #' df <- oa2df(res, entity = "concepts")
@@ -588,6 +532,37 @@ concepts2df <- function(data, verbose = TRUE,
 
   out_df <- rbind_oa_ls(list_df)
   out_df[, intersect(col_order, names(out_df))]
+}
+
+
+#' Convert keywords from list to data frame
+#'
+#' The function converts a list of keywords obtained using \code{oa_request} or
+#' \code{oa_fetch(output = "list")} into a data frame/tibble. More on keyword at
+#' <https://help.openalex.org/hc/en-us/articles/24736201130391-Keywords>.
+#'
+#' @inheritParams works2df
+#'
+#' @return a data.frame.
+#'
+#'
+#' @examples
+#' \dontrun{
+#'
+#' x <- oa_fetch(
+#'   entity = "keywords",
+#'   options = list(sample = 5),
+#'   output = "list"
+#' )
+#'
+#' df <- oa2df(x, entity = "keywords")
+#'
+#' df
+#' }
+#'
+#' @export
+keywords2df <- function(data, verbose = TRUE) {
+  tibble::as_tibble(subs_na(data, "rbind_df")[[1]])
 }
 
 
@@ -708,6 +683,7 @@ sources2df <- function(data, verbose = TRUE,
     "flat", "summary_stats",
     "identical", "is_oa",
     "identical", "is_in_doaj",
+    "identical", "is_indexed_in_scopus",
     "flat", "ids",
     "identical", "homepage_url",
     "rbind_df", "apc_prices",
@@ -839,15 +815,10 @@ publishers2df <- function(data, verbose = TRUE,
 #' # Query to search information about all Italian educational institutions
 #'
 #'
-#' query_inst <- oa_query(
+#' res <- oa_query(
 #'   entity = "topics",
-#'   display_name.search = "electrodynamics"
-#' )
-#'
-#' res <- oa_request(
-#'   query_url = query_inst,
-#'   count_only = FALSE,
-#'   verbose = FALSE
+#'   display_name.search = "electrodynamics",
+#'   output = "list"
 #' )
 #'
 #' df <- oa2df(res, entity = "topics")
@@ -907,11 +878,6 @@ topics2df <- function(data, verbose = TRUE,
 
 
 #' Flatten snowball result
-#'
-#' |  id|title |...|cited_by_count| referenced_works   |cited_by |...|
-#' | 100|foo   |...|             1| 98, 99             |101      |...|
-#' | 200|bar   |...|             2| 198, 199           |201, 202 |...|
-#' | 300|wug   |...|             2| 296, 297, 298, 299 |301, 302 |...|
 #'
 #' @param data List result from `oa_snowball`.
 #' @param verbose Logical. If TRUE, print information on wrangling process.
@@ -986,5 +952,5 @@ snowball2df <- function(data, verbose = FALSE) {
   nodes_augmented$connection_count <-
     nodes_augmented$backward_count + nodes_augmented$forward_count
 
-  nodes_augmented
+  tibble::as_tibble(nodes_augmented)
 }