+++ NEWS.md
@@ -1,3 +1,27 @@
+# openalexR 2.0.0
+
+# openalexR 2.0.0
+* Breaking changes in column names in the output of `oa_fetch`:
+  * Works:
+    * `so` is now `source_display_name`
+    * `so_id` is now `source_id`
+    * `host_organization` now contains the id of the host organization
+    * `host_organization_name` now contains the name of the host organization
+    * `ab` is now `abstract`
+    * `url` is now `landing_page_url`
+    * `author` is now `authorships`
+    * the nested columns under `authors` no longer have the `au` prefix
+    * New columns: `fwci`, `referenced_works_count`, `keywords`
+  * Authors:
+    * `affiliation*` is removed
+    * New columns: `last_known_institutions`, `2yr_mean_citedness`, `h_index`, `i10_index`
+* Removed `concepts` as an entity
+* Added `keywords` as an entity
+* Added `get_coverage()` to track the oa2df-mapped columns of OpenAlex fields
+  
+* Deprecated `oa2bibliometrix()`. Use `bibliometrix::convert2df()` 
+(from the **bibliometrix** R package) instead.
+
 # openalexR 1.4.0
 * "topics" are now a valid entity in oa_fetch
 * The column "topics" replaces concepts in most entities' returned dataframes